Split extensions out and comment a bit
